Extract PipelineJobIdUtils
