+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/api/job/PipelineJobId.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.data.pipeline.api.job;
 
 /**
- * Job id.
+ * Pipeline job id.
  */
 public interface PipelineJobId {
